improve logging names and package structure
