+++ pkg/logging/logger/elasticsearch/es_logger.go
@@ -24,7 +24,7 @@ import (
 	"net/http"
 
 	"github.com/elastic/cloud-sdk-go/pkg/logging"
-	standardoutputlogger "github.com/elastic/cloud-sdk-go/pkg/logging/standardoutput"
+	logger "github.com/elastic/cloud-sdk-go/pkg/logging/logger/standardoutput"
 )
 
 const (
@@ -38,17 +38,17 @@ type ESLogger struct {
 
 // Log sends a logging message to an ES cluster
 // If logger fails to deliver the message then logs an error message to the standard output logger
-func (logger ESLogger) Log(msg logging.LogMessage) error {
+func (esLogger ESLogger) Log(msg logging.LogMessage) error {
 	b, err := msg.Marshall()
 	if err != nil {
 		return err
 	}
-	req, err := http.NewRequest(http.MethodPost, fmt.Sprintf("%s%s", logger.host, logIndex), bytes.NewBuffer(b))
+	req, err := http.NewRequest(http.MethodPost, fmt.Sprintf("%s%s", esLogger.host, logIndex), bytes.NewBuffer(b))
 	if err != nil {
 		return err
 	}
 
-	req.SetBasicAuth(logger.user, logger.pass)
+	req.SetBasicAuth(esLogger.user, esLogger.pass)
 	req.Header.Set("Content-Type", "application/json")
 
 	client := &http.Client{}
@@ -67,34 +67,34 @@ func (logger ESLogger) Log(msg logging.LogMessage) error {
 				logging.NewAgent().WithName("cloud-sdk-go")).
 			WithMessage(fmt.Sprintf(
 				"cannot send logs to ES cluster %s. error code is %d\n detailed error %s",
-				logger.host,
+				esLogger.host,
 				resp.StatusCode,
 				string(body)))
 
 		// Do we want to check the error here?
 		// nolint
-		logging.NewLogDispatcher(standardoutputlogger.New()).Dispatch(msg)
+		logger.NewStandardOutputDispatcher().Log(msg)
 	}
 
 	return nil
 }
 
 // WithHost sets the ES host and returns the ESLogger itself
-func (logger ESLogger) WithHost(host string) ESLogger {
-	logger.host = host
-	return logger
+func (esLogger ESLogger) WithHost(host string) ESLogger {
+	esLogger.host = host
+	return esLogger
 }
 
 // WithUser sets the ES user and returns the ESLogger itself
-func (logger ESLogger) WithUser(user string) ESLogger {
-	logger.user = user
-	return logger
+func (esLogger ESLogger) WithUser(user string) ESLogger {
+	esLogger.user = user
+	return esLogger
 }
 
 // WithPass sets the ES password and returns the ESLogger itself
-func (logger ESLogger) WithPass(pass string) ESLogger {
-	logger.pass = pass
-	return logger
+func (esLogger ESLogger) WithPass(pass string) ESLogger {
+	esLogger.pass = pass
+	return esLogger
 }
 
 // New properly creates a new standard output logger initializing its internal state with default values
