+++ pkg/logging/logging.go
@@ -18,9 +18,13 @@
 package logging
 
 import (
+	"strings"
 	"sync"
 
+	"github.com/fatih/color"
 	multierror "github.com/hashicorp/go-multierror"
+
+	loggingdecorator "github.com/elastic/cloud-sdk-go/pkg/logging/decorator"
 )
 
 // LogLevel defines the logging level as untyped int
@@ -37,6 +41,12 @@ const (
 	DEBUG
 	// TRACE defines the trace log level as int value of 5
 	TRACE
+
+	warn  = "WARNING"
+	trace = "TRACE"
+	info  = "INFO"
+	err   = "ERROR"
+	debug = "DEBUG"
 )
 
 // Logger is the interface describing a loggers's behavior
@@ -92,4 +102,76 @@ func (ld *MuxDispatcher) Add(loggers ...Logger) {
 	ld.lock.Lock()
 	ld.loggers = append(ld.loggers, loggers...)
 	ld.lock.Unlock()
+}
+
+// LevelColor returns the color attribute based on the log level and the coloring scheme
+func LevelColor(scheme loggingdecorator.ColoringScheme, logLevel string) color.Attribute {
+	switch strings.ToUpper(logLevel) {
+	case trace:
+		return scheme.Trace
+	case debug:
+		return scheme.Debug
+	case info:
+		return scheme.Info
+	case warn:
+		return scheme.Warn
+	case err:
+		return scheme.Error
+	default:
+		return color.FgWhite
+	}
+}
+
+// WarnMessage creates a new WARNING log message and properly initializes all required fields
+func WarnMessage(msg string) LogMessage {
+	return newLogMessage(WARNING).WithMessage(msg)
+}
+
+// ErrorMessage creates a new ERROR log message and properly initializes all required fields
+func ErrorMessage(msg string) LogMessage {
+	return newLogMessage(ERROR).WithMessage(msg).WithError(NewError().WithMessage(msg))
+}
+
+// InfoMessage creates a new INFO log message and properly initializes all required fields
+func InfoMessage(msg string) LogMessage {
+	return newLogMessage(INFO).WithMessage(msg)
+}
+
+// DebugMessage creates a new DEBUG log message and properly initializes all required fields
+func DebugMessage(msg string) LogMessage {
+	return newLogMessage(DEBUG).WithMessage(msg)
+}
+
+// TraceMessage creates a new TRACE log message and properly initializes all required fields
+func TraceMessage(msg string) LogMessage {
+	return newLogMessage(TRACE).WithMessage(msg)
+}
+
+func newLogMessage(level LogLevel) LogMessage {
+	return NewLogMessage().
+		WithAgent(NewAgent()).
+		WithLog(NewLog().
+			WithLevel(level))
+}
+
+// Level returns the LogLevel based on the given level as string
+// If the level is not found then the default (INFO) is returned
+func Level(level string) LogLevel {
+	var logLevel LogLevel
+	switch level {
+	case trace:
+		logLevel = TRACE
+	case debug:
+		logLevel = DEBUG
+	case info:
+		logLevel = INFO
+	case warn:
+		logLevel = WARNING
+	case err:
+		logLevel = ERROR
+	default:
+		logLevel = INFO
+	}
+
+	return logLevel
 }
